# backend/api/report/v1/report_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v3.21.12
// source: report/v1/report.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationReportDownloadExport = "/report.v1.Report/DownloadExport"
const OperationReportExportUserFeedback = "/report.v1.Report/ExportUserFeedback"
const OperationReportGetExportStatus = "/report.v1.Report/GetExportStatus"

type ReportHTTPServer interface {
	// DownloadExport Download exported file
	DownloadExport(context.Context, *DownloadExportRequest) (*DownloadExportReply, error)
	// ExportUserFeedback Export user feedback data as Excel
	ExportUserFeedback(context.Context, *ExportUserFeedbackRequest) (*ExportUserFeedbackReply, error)
	// GetExportStatus Get export status
	GetExportStatus(context.Context, *GetExportStatusRequest) (*GetExportStatusReply, error)
}

func RegisterReportHTTPServer(s *http.Server, srv ReportHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/report/export/userfeedback", _Report_ExportUserFeedback0_HTTP_Handler(srv))
	r.GET("/v1/report/export/status/{export_id}", _Report_GetExportStatus0_HTTP_Handler(srv))
	r.GET("/v1/report/export/download/{export_id}", _Report_DownloadExport0_HTTP_Handler(srv))
}

func _Report_ExportUserFeedback0_HTTP_Handler(srv ReportHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ExportUserFeedbackRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationReportExportUserFeedback)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ExportUserFeedback(ctx, req.(*ExportUserFeedbackRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ExportUserFeedbackReply)
		return ctx.Result(200, reply)
	}
}

func _Report_GetExportStatus0_HTTP_Handler(srv ReportHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetExportStatusRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationReportGetExportStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetExportStatus(ctx, req.(*GetExportStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetExportStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Report_DownloadExport0_HTTP_Handler(srv ReportHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DownloadExportRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationReportDownloadExport)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DownloadExport(ctx, req.(*DownloadExportRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DownloadExportReply)
		return ctx.Result(200, reply)
	}
}

type ReportHTTPClient interface {
	// DownloadExport Download exported file
	DownloadExport(ctx context.Context, req *DownloadExportRequest, opts ...http.CallOption) (rsp *DownloadExportReply, err error)
	// ExportUserFeedback Export user feedback data as Excel
	ExportUserFeedback(ctx context.Context, req *ExportUserFeedbackRequest, opts ...http.CallOption) (rsp *ExportUserFeedbackReply, err error)
	// GetExportStatus Get export status
	GetExportStatus(ctx context.Context, req *GetExportStatusRequest, opts ...http.CallOption) (rsp *GetExportStatusReply, err error)
}

type ReportHTTPClientImpl struct {
	cc *http.Client
}

func NewReportHTTPClient(client *http.Client) ReportHTTPClient {
	return &ReportHTTPClientImpl{client}
}

// DownloadExport Download exported file
func (c *ReportHTTPClientImpl) DownloadExport(ctx context.Context, in *DownloadExportRequest, opts ...http.CallOption) (*DownloadExportReply, error) {
	var out DownloadExportReply
	pattern := "/v1/report/export/download/{export_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationReportDownloadExport))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ExportUserFeedback Export user feedback data as Excel
func (c *ReportHTTPClientImpl) ExportUserFeedback(ctx context.Context, in *ExportUserFeedbackRequest, opts ...http.CallOption) (*ExportUserFeedbackReply, error) {
	var out ExportUserFeedbackReply
	pattern := "/v1/report/export/userfeedback"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationReportExportUserFeedback))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetExportStatus Get export status
func (c *ReportHTTPClientImpl) GetExportStatus(ctx context.Context, in *GetExportStatusRequest, opts ...http.CallOption) (*GetExportStatusReply, error) {
	var out GetExportStatusReply
	pattern := "/v1/report/export/status/{export_id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationReportGetExportStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}