# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/chat/message:
        post:
            tags:
                - Chat
            description: Send a chat message
            operationId: Chat_SendMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/chat.v1.SendMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/chat.v1.SendMessageReply'
    /v1/chat/messages:
        get:
            tags:
                - Chat
            description: Get chat messages
            operationId: Chat_GetMessages
            parameters:
                - name: sessionId
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/chat.v1.GetMessagesReply'
...
    /v1/report/export/download/{exportId}:
        get:
            tags:
                - Report
            description: Download exported file
            operationId: Report_DownloadExport
            parameters:
                - name: exportId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/report.v1.DownloadExportReply'
    /v1/report/export/status/{exportId}:
        get:
            tags:
                - Report
            description: Get export status
            operationId: Report_GetExportStatus
            parameters:
                - name: exportId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/report.v1.GetExportStatusReply'
    /v1/report/export/userfeedback:
        post:
            tags:
                - Report
            description: Export user feedback data as Excel
            operationId: Report_ExportUserFeedback
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/report.v1.ExportUserFeedbackRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/report.v1.ExportUserFeedbackReply'
...
components:
    schemas:
        chat.v1.ChatMessage:
            type: object
            properties:
                id:
                    type: string
                content:
                    type: string
                userId:
                    type: string
                sessionId:
                    type: string
                timestamp:
                    type: string
                type:
                    type: string
            description: Chat message structure
        chat.v1.GetMessagesReply:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/chat.v1.ChatMessage'
                total:
                    type: integer
                    format: int32
            description: Reply for getting messages
        chat.v1.SendMessageReply:
            type: object
            properties:
                messageId:
                    type: string
                success:
                    type: boolean
                errorMessage:
                    type: string
            description: Reply for sending a message
        chat.v1.SendMessageRequest:
            type: object
            properties:
                content:
                    type: string
                userId:
                    type: string
                sessionId:
                    type: string
            description: Request for sending a message
...
        report.v1.DownloadExportReply:
            type: object
            properties:
                fileContent:
                    type: string
                    format: bytes
                fileName:
                    type: string
                contentType:
                    type: string
            description: Reply for downloading export
        report.v1.ExportUserFeedbackReply:
            type: object
            properties:
                exportId:
                    type: string
                status:
                    type: string
                message:
                    type: string
            description: Reply for exporting user feedback
        report.v1.ExportUserFeedbackRequest:
            type: object
            properties:
                startDate:
                    type: string
                endDate:
                    type: string
                filters:
                    type: array
                    items:
                        type: string
                userId:
                    type: string
            description: Request for exporting user feedback
        report.v1.GetExportStatusReply:
            type: object
            properties:
                exportId:
                    type: string
                status:
                    type: string
                fileUrl:
                    type: string
                errorMessage:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: Reply for getting export status
...